+++ lib/moderation/storage/in_memory.rb
@@ -5,8 +5,8 @@ module Moderation
     class InMemory < Moderation::AbstractStorage
       attr_accessor :limit
 
-      def initialize(limit = Moderation::Store::DEFAULT_LIMIT)
-        @limit = limit
+      def initialize(options = {})
+        @limit = options.fetch(:limit, Moderation::Store::DEFAULT_LIMIT)
       end
 
       def insert(item)
@@ -20,6 +20,10 @@ module Moderation
       def all(options = {})
         fetch_limit = options.fetch(:limit, limit)
         data.first(fetch_limit)
+      end
+
+      def clean!
+        @data = []
       end
 
       private